MLOps часть/frontend/requirements.txt
streamlit==1.10.0
pandas~=1.3.5
requests~=2.27.1
pyyaml==6.0
joblib~=1.1.0
numpy~=1.20.3
optuna~=2.10.0
matplotlib==3.5.0
seaborn==0.11.2
plotly==5.3.1
scikit-learn~=1.0.2
